Database fixed
An error in Client Table

+++ Entity/ClientEntity.cs
@@ -9,14 +9,14 @@ namespace Entity
     public class ClientEntity
     {
 
-        private int dni;
+        private long dni;
         private string name;
         private string surname;
         private DateTime licenseValidDate;
         private int quantityOfRents;
         private bool blackList;
 
-        public int Dni { get => dni; set => dni = value; }
+        public long Dni { get => dni; set => dni = value; }
         public string Name { get => name; set => name = value; }
         public string Surname { get => surname; set => surname = value; }
         public DateTime LicenseValidDate { get => licenseValidDate; set => licenseValidDate = value; }
